+++ cli/commands/project_sync.go
@@ -39,8 +39,24 @@ func (cmd *ProjectSync) Commands() []cli.Command {
 		Usage:       "Start a unison sync on local project directory. Optionally provide a volume name.",
 		ArgsUsage:   "[optional volume name]",
 		Description: "Volume name will be discovered in the following order: argument to this command > outrigger project config > docker-compose file > current directory name",
-		Before:      cmd.Before,
-		Action:      cmd.RunStart,
+		Flags: []cli.Flag{
+			cli.IntFlag{
+				Name:   "initial-sync-timeout",
+				Value:  60,
+				Usage:  "Maximum amount of time in seconds to allow for detecting each of start of the unison container and start of initial sync",
+				EnvVar: "RIG_PROJECT_SYNC_TIMEOUT",
+			},
+			// Arbitrary sleep length but anything less than 3 wasn't catching
+			// ongoing very quick file updates during a test
+			cli.IntFlag{
+				Name:   "initial-sync-wait",
+				Value:  5,
+				Usage:  "Time in seconds to wait between checks to see if initial sync has finished.",
+				EnvVar: "RIG_PROJECT_INITIAL_SYNC_WAIT",
+			},
+		},
+		Before: cmd.Before,
+		Action: cmd.RunStart,
 	}
 	stop := cli.Command{
 		Name:        "sync:stop",
@@ -84,7 +100,7 @@ func (cmd *ProjectSync) RunStart(ctx *cli.Context) error {
 		cmd.out.Error.Fatalf("Error starting sync container %s: %v", volumeName, err)
 	}
 
-	var ip = cmd.WaitForUnisonContainer(volumeName)
+	var ip = cmd.WaitForUnisonContainer(volumeName, ctx.Int("initial-sync-timeout"))
 
 	cmd.out.Info.Println("Initializing sync")
 
@@ -113,7 +129,7 @@ func (cmd *ProjectSync) RunStart(ctx *cli.Context) error {
 		cmd.out.Error.Fatalf("Error starting local unison process: %v", err)
 	}
 
-	cmd.WaitForSyncInit(logFile)
+	cmd.WaitForSyncInit(logFile, ctx.Int("initial-sync-timeout"), ctx.Int("initial-sync-wait"))
 
 	return nil
 }
@@ -183,8 +199,12 @@ func (cmd *ProjectSync) LoadComposeFile() (*ComposeFile, error) {
 // we need to discover the IP address of the container instead of using the DNS name
 // when compiled without -cgo this executable will not use the native mac dns resolution
 // which is how we have configured dnsdock to provide names for containers.
-func (cmd *ProjectSync) WaitForUnisonContainer(containerName string) string {
+func (cmd *ProjectSync) WaitForUnisonContainer(containerName string, timeoutSeconds int) string {
 	cmd.out.Info.Println("Waiting for container to start")
+
+	var timeoutLoopSleep = time.Duration(100) * time.Millisecond
+	// * 10 here because we loop once every 100 ms and we want to get to seconds
+	var timeoutLoops = timeoutSeconds * 10
 
 	output, err := exec.Command("docker", "inspect", "--format", "{{.NetworkSettings.IPAddress}}", containerName).Output()
 	if err != nil {
@@ -193,13 +213,13 @@ func (cmd *ProjectSync) WaitForUnisonContainer(containerName string) string {
 	ip := strings.Trim(string(output), "\n")
 
 	cmd.out.Verbose.Printf("Checking for unison network connection on %s %d", ip, UNISON_PORT)
-	for i := 1; i <= 100; i++ {
+	for i := 1; i <= timeoutLoops; i++ {
 		if conn, err := net.Dial("tcp", fmt.Sprintf("%s:%d", ip, UNISON_PORT)); err == nil {
 			conn.Close()
 			return ip
 		} else {
 			cmd.out.Info.Printf("Error: %v", err)
-			time.Sleep(time.Duration(100) * time.Millisecond)
+			time.Sleep(timeoutLoopSleep)
 		}
 	}
 	cmd.out.Error.Fatal("Sync container failed to start!")
@@ -207,32 +227,51 @@ func (cmd *ProjectSync) WaitForUnisonContainer(containerName string) string {
 }
 
 // The local unison process is finished initializing when the log file exists
-func (cmd *ProjectSync) WaitForSyncInit(logFile string) {
-	cmd.out.Info.Print("Waiting for initial sync to finish...")
+// and has stopped growing in size
+func (cmd *ProjectSync) WaitForSyncInit(logFile string, timeoutSeconds int, syncWaitSeconds int) {
+	cmd.out.Info.Print("Waiting for initial sync detection")
 
 	var tempFile = fmt.Sprintf(".%s.tmp", logFile)
+	var timeoutLoopSleep = time.Duration(100) * time.Millisecond
+	// * 10 here because we loop once every 100 ms and we want to get to seconds
+	var timeoutLoops = timeoutSeconds * 10
 
 	// Create a temp file to cause a sync action
 	exec.Command("touch", tempFile).Run()
 
-	// Lets check for 60 seconds, while waiting for initial sync to complete
-	for i := 1; i <= 600; i++ {
+	for i := 1; i <= timeoutLoops; i++ {
 		if i%10 == 0 {
 			os.Stdout.WriteString(".")
 		}
-		if _, err := os.Stat(logFile); err == nil {
-			// Remove the temp file now that we are running
-			os.Stdout.WriteString("done\n")
+		if statInfo, err := os.Stat(logFile); err == nil {
+			os.Stdout.WriteString(" initial sync detected\n")
+
+			cmd.out.Info.Print("Waiting for initial sync to finish")
+			var statSleep = time.Duration(syncWaitSeconds) * time.Second
+			// Initialize at -2 to force at least one loop
+			var lastSize = int64(-2)
+			for lastSize != statInfo.Size() {
+				os.Stdout.WriteString(".")
+				time.Sleep(statSleep)
+				lastSize = statInfo.Size()
+				if statInfo, err = os.Stat(logFile); err != nil {
+					cmd.out.Info.Print(err.Error())
+					lastSize = -1
+				}
+			}
+			os.Stdout.WriteString(" done\n")
+			// Remove the temp file, waiting until after sync so spurious
+			// failure message doesn't show in log
 			exec.Command("rm", "-f", tempFile).Run()
 			return
 		} else {
-			time.Sleep(time.Duration(100) * time.Millisecond)
+			time.Sleep(timeoutLoopSleep)
 		}
 	}
 
 	// The log file was not created, the sync has not started yet
 	exec.Command("rm", "-f", tempFile).Run()
-	cmd.out.Error.Fatal("Sync container failed to start!")
+	cmd.out.Error.Fatal("Failed to detect start of initial sync!")
 }
 
 // Get the local Unison version to try to load a compatible unison image